Add security helper tests and refactor auth utilities
